✨ Sow-able grass, growable trees!
BehaviourObj is now a base-class for crafting behaviours.

Can craft in 2 ways:
- World Interaction
- Inventory to Item in World

Next is to do the in-inventory crafting. Hold an item then try hold another item. It might not be necessary, but it might not also be too hard to do.

I'll do bugs next. These are obtained through interacting with grass.